+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 9.1.0-dev
 - allowed setting base url for Nette\Application\LinkGenerator from Nette DIC extension
 - added arguments --bootstrap and --list-test-suites for mytester.php
 - added option to use a config file for mytester.php
+- allowed using methods without test prefix as test methods in TestCase
 
 Version 9.0.2
 - fixed parameter $config of ContainerFactory::create()/TCompiledContainer::refreshContainer() not working